f1-companion/backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-SocketIO
Flask-Caching
requests
python-dotenv
psycopg2-binary
Flask-Migrate